Move default entropy function code from platform.c to platform_util.c
This is so that it is no longer guarded by MBEDTLS_PLATFORM_C
